requirements.txt
evaluate==0.4.3
Flask==3.0.3
matplotlib==3.8.0
numpy==1.23.5
opencv-python==4.10.0.84
pillow==11.0.0
PyYAML==6.0.2
scikit-learn==1.5.2
torch==2.5.0
torchvision==0.20.0
transformers==4.46.2
tqdm==4.66.6